requirements.txt
Scrapy
python_dateutil
Flask
python-dateutil
PyYAML
pika